Make nanoseconds_t a double to prevent overflow on very large numbers of revolutions (I've just seen a flux file with 50).

+++ lib/globals.h
@@ -12,7 +12,7 @@
 #include <set>
 #include <cassert>
 
-typedef int nanoseconds_t;
+typedef double nanoseconds_t;
 class Bytes;
 
 extern double getCurrentTime();
